WordCamp Central: Add year to camp dates on the schedule
Changes the Schedule page template to emulate the organization on the
Past Camps page template, by grouping camps under a year header and
including the year in each camp's date.

This also adds a param to `WordCamp_Central_Theme::the_wordcamp_date` to
include the year in the camp date output. The default is to not include
it, for backcompat.

Props frq
Fixes #2065

git-svn-id: https://meta.svn.wordpress.org/sites/trunk/wordcamp.org/public_html/wp-content/themes/wordcamp-central-2012@5530 74240141-8908-4e6f-9713-ba540dce6ec7